give user option to update the indiserver

+++ setup.sh
@@ -3,6 +3,7 @@
 #set -x  # command tracing
 #set -o errexit  # replace by trapping ERR
 #set -o nounset  # problems with python virtualenvs
+shopt -s nullglob
 
 PATH=/usr/bin:/bin
 export PATH
@@ -11,7 +12,7 @@ export PATH
 #### config ####
 INDI_ALLSKY_VERSION="20221023.0"
 INDI_DRIVER_PATH="/usr/bin"
-INDISEVER_SERVICE_NAME="indiserver"
+INDISERVER_SERVICE_NAME="indiserver"
 ALLSKY_SERVICE_NAME="indi-allsky"
 GUNICORN_SERVICE_NAME="gunicorn-indi-allsky"
 ALLSKY_ETC="/etc/indi-allsky"
@@ -22,6 +23,7 @@ DB_FILE="${DB_FOLDER}/indi-allsky.sqlite"
 DB_URI_DEFAULT="sqlite:///${DB_FILE}"
 INSTALL_INDI="true"
 INSTALL_LIBCAMERA="false"
+INSTALL_INDISERVER="true"
 HTTP_PORT="80"
 HTTPS_PORT="443"
 DPC_STRENGTH="0"
@@ -35,7 +37,6 @@ DPC_STRENGTH="0"
 # 1 = Normal correction (default)
 # 2 = Strong correction
 ###
-
 
 
 function catch_error() {
@@ -140,7 +141,7 @@ echo "Release: $DISTRO_RELEASE"
 echo "Arch: $CPU_ARCH"
 echo
 echo "INDI_DRIVER_PATH: $INDI_DRIVER_PATH"
-echo "INDISERVER_SERVICE_NAME: $INDISEVER_SERVICE_NAME"
+echo "INDISERVER_SERVICE_NAME: $INDISERVER_SERVICE_NAME"
 echo "ALLSKY_SERVICE_NAME: $ALLSKY_SERVICE_NAME"
 echo "GUNICORN_SERVICE_NAME: $GUNICORN_SERVICE_NAME"
 echo "ALLSKY_ETC: $ALLSKY_ETC"
@@ -187,6 +188,8 @@ echo "indi-allsky supports the following camera interfaces."
 echo
 echo "Note:  libcamera is generally only available on ARM SoCs like Raspberry Pi"
 echo
+
+# whiptail might not be installed yet
 PS3="Select a camera interface: "
 select camera_interface in indi libcamera_imx477; do
     if [ -n "$camera_interface" ]; then
@@ -1056,11 +1059,17 @@ if [ -z "$DBUS_SESSION_BUS_ADDRESS" ]; then
 fi
 
 
-
-# get list of drivers
-cd "$INDI_DRIVER_PATH"
-INDI_DRIVERS=$(ls indi_*_ccd indi_rpicam 2>/dev/null || true)
-cd "$OLDPWD"
+if systemctl -q is-enabled "${INDISERVER_SERVICE_NAME}" 2>/dev/null; then
+    # system
+    INSTALL_INDISERVER="no"
+elif systemctl --user -q is-enabled "${INDISERVER_SERVICE_NAME}" 2>/dev/null; then
+    # user
+    if whiptail --title "indiserver update" --yesno "An indiserver service is already defined, would you like to replace it?" 0 0 --defaultno; then
+        INSTALL_INDISERVER="true"
+    else
+        INSTALL_INDISERVER="false"
+    fi
+fi
 
 
 # find script directory for service setup
@@ -1098,15 +1107,22 @@ pip3 install --upgrade pip setuptools wheel
 pip3 install -r "${ALLSKY_DIRECTORY}/${VIRTUALENV_REQ}"
 
 
-if [ "$CAMERA_INTERFACE" == "indi" ]; then
-    echo
-    echo
-    PS3="Select an INDI driver: "
-    select indi_driver_path in $INDI_DRIVERS; do
-        if [ -f "${INDI_DRIVER_PATH}/${indi_driver_path}" ]; then
-            CCD_DRIVER=$indi_driver_path
-            break
-        fi
+
+# get list of drivers
+INDI_DRIVERS=""
+cd "$INDI_DRIVER_PATH"
+for I in indi_*_ccd indi_rpicam*; do
+    INDI_DRIVERS="$INDI_DRIVERS $I $I OFF "
+done
+cd "$OLDPWD"
+
+#echo $INDI_DRIVERS
+
+
+CCD_DRIVER=""
+if [[ "$CAMERA_INTERFACE" == "indi" && "$INSTALL_INDISERVER" == "true" ]]; then
+    while [ -z "$CCD_DRIVER" ]; do
+        CCD_DRIVER=$(whiptail --title "Camera Driver" --nocancel --notags --radiolist "Press space to select" 0 0 0 $INDI_DRIVERS 3>&1 1>&2 2>&3)
     done
 else
     # simulator will not affect anything
@@ -1119,20 +1135,28 @@ fi
 # create users systemd folder
 [[ ! -d "${HOME}/.config/systemd/user" ]] && mkdir -p "${HOME}/.config/systemd/user"
 
-echo
-echo
-echo "**** Setting up indiserver service ****"
-TMP1=$(mktemp)
-sed \
- -e "s|%INDI_DRIVER_PATH%|$INDI_DRIVER_PATH|g" \
- -e "s|%INDISERVER_USER%|$USER|g" \
- -e "s|%INDI_CCD_DRIVER%|$CCD_DRIVER|g" \
- ${ALLSKY_DIRECTORY}/service/indiserver.service > $TMP1
+
+if [ "$INSTALL_INDISERVER" == "true" ]; then
+    echo
+    echo
+    echo "**** Setting up indiserver service ****"
+    TMP1=$(mktemp)
+    sed \
+     -e "s|%INDI_DRIVER_PATH%|$INDI_DRIVER_PATH|g" \
+     -e "s|%INDISERVER_USER%|$USER|g" \
+     -e "s|%INDI_CCD_DRIVER%|$CCD_DRIVER|g" \
+     ${ALLSKY_DIRECTORY}/service/indiserver.service > $TMP1
 
 
-cp -f "$TMP1" "${HOME}/.config/systemd/user/${INDISEVER_SERVICE_NAME}.service"
-chmod 644 "${HOME}/.config/systemd/user/${INDISEVER_SERVICE_NAME}.service"
-[[ -f "$TMP1" ]] && rm -f "$TMP1"
+    cp -f "$TMP1" "${HOME}/.config/systemd/user/${INDISERVER_SERVICE_NAME}.service"
+    chmod 644 "${HOME}/.config/systemd/user/${INDISERVER_SERVICE_NAME}.service"
+    [[ -f "$TMP1" ]] && rm -f "$TMP1"
+else
+    echo
+    echo
+    echo
+    echo "! Bypassing indiserver setup"
+fi
 
 
 echo "**** Setting up indi-allsky service ****"
@@ -1175,7 +1199,7 @@ chmod 644 "${HOME}/.config/systemd/user/${GUNICORN_SERVICE_NAME}.service"
 echo "**** Enabling services ****"
 sudo loginctl enable-linger $USER
 systemctl --user daemon-reload
-systemctl --user enable ${INDISEVER_SERVICE_NAME}.service
+systemctl --user enable ${INDISERVER_SERVICE_NAME}.service
 systemctl --user enable ${ALLSKY_SERVICE_NAME}.service
 systemctl --user enable ${GUNICORN_SERVICE_NAME}.socket
 systemctl --user enable ${GUNICORN_SERVICE_NAME}.service
@@ -1294,7 +1318,7 @@ sed \
  -e "s|%ALLSKY_ETC%|$ALLSKY_ETC|g" \
  -e "s|%HTDOCS_FOLDER%|$HTDOCS_FOLDER|g" \
  -e "s|%IMAGE_FOLDER%|$IMAGE_FOLDER|g" \
- -e "s|%INDISEVER_SERVICE_NAME%|$INDISEVER_SERVICE_NAME|g" \
+ -e "s|%INDISERVER_SERVICE_NAME%|$INDISERVER_SERVICE_NAME|g" \
  -e "s|%ALLSKY_SERVICE_NAME%|$ALLSKY_SERVICE_NAME|g" \
  -e "s|%GUNICORN_SERVICE_NAME%|$GUNICORN_SERVICE_NAME|g" \
  "${ALLSKY_DIRECTORY}/flask.json_template" > $TMP4
